ogc api contracts
