+++ source/WaDEApiFunctions/Program.cs
@@ -72,4 +72,4 @@ var host = new HostBuilder()
     })
     .Build();
 
-host.Run();
+await host.RunAsync();
