+++ source/WaDEApiFunctions/local.settings.json
@@ -3,6 +3,7 @@
   "Values": {
     "AzureWebJobsStorage": "UseDevelopmentStorage=true",
     "FUNCTIONS_WORKER_RUNTIME": "dotnet-isolated",
+    "OpenApi__HostNames": "http://localhost:7071/api",
     "OpenApi__Version": "v3",
     "OpenApi__DocTitle": "WaDE API",
     "OpenApi__ForceHttps": "false",
