+++ source/WesternStatesWater.WaDE.Contracts.Api/OgcApi/Collection.cs
@@ -1,0 +1,12 @@
+namespace WesternStatesWater.WaDE.Contracts.Api.OgcApi;
+
+public class Collection
+{
+    public required string Id { get; set; }
+    public required Link[] Links { get; set; }
+    public long StorageCrsCoordinateEpoch { get; set; }
+    public string[] Crs { get; set; }
+    public string ItemType { get; set; }
+    public string StorageCrs { get; set; }
+    public Extent Extent { get; set; }
+}
