+++ source/WesternStatesWater.WaDE.Contracts.Api/OgcApi/Extent.cs
@@ -1,0 +1,7 @@
+namespace WesternStatesWater.WaDE.Contracts.Api.OgcApi;
+
+public class Extent
+{
+    public Spatial Spatial { get; set; }
+    public Temporal Temporal { get; set; }
+}
