+++ source/WesternStatesWater.WaDE.Contracts.Api/OgcApi/Spatial.cs
@@ -1,0 +1,7 @@
+namespace WesternStatesWater.WaDE.Contracts.Api.OgcApi;
+
+public class Spatial
+{
+    public string? Crs { get; set; }
+    public double[][] Bbox { get; set; }
+}
