+++ source/WesternStatesWater.WaDE.Contracts.Api/OgcApi/Temporal.cs
@@ -1,0 +1,7 @@
+namespace WesternStatesWater.WaDE.Contracts.Api.OgcApi;
+
+public class Temporal
+{
+    public string[][] Interval { get; set; }
+    public string Trs { get; set; }
+}
